+++ aks_node_pool/provider.tf
@@ -1,4 +1,4 @@
 provider "azurerm" {
-  version = "~> 2.5.0"
+  version = "~> 2.16.0"
   features {}
 }